Rework testing again.
